Adds a customized template for API controllers.
This customized template adds more comments to the code to help students
understand what all the parts of the code are doing. This would be too
much for a regular API controller but may help students better
understand the generated code when they need to modify any of it for
their needs.